Validation demonstration
First we put some rubbish in the field and we see that our black box is
failing. The following should restart the application three times which
is a good opportunity to show how boring it is.

First, we make sure that the version is valid by adding a constraint
validation. We need to move the two fields in a DiffRequest nested class
and flag the method parameter in the controller with `@Valid` and
`@ModelAttributes`. If we restart, we can see that the rubbish value is
rejected now.

Then, we show that if we put a more recent version as the original
version, nothing happens really. We then upgrade the loader to run
the `VersionStringComparator`. We should also annotate the exception so
that it returns a 400 instead of 500.

But that's still showing the default (white label) error page. Let's
customize the container with a custom 400 error page.