Devtools demonstration
Various devtools demonstration. After adding the dependency
and starting the project we can connect the browser to the LiveReload
server.

We can diff 1.2.5.RELEASE and 1.3.0.M3 that should show that the
`spring.view.*` keys are deprecated. We don't see it because we're
missing a CSS entry. Let's add it

```
.deprecated {
    text-decoration: line-through
}
```

Then, it would be nice if we could filter out the groups that have not
changed. For that we need to change the template to offer an additional
option (a checkbox). And we need to handle the new boolean value on
the server-side, filtering the collection if necessary.

Finally, if we go back on the home page, we have an empty form. It would
be nice if it was pre-filled with some data. Let's update the controller
to set some values in there. It should show up immediately in the browser
since LiveReload makes sure to refresh the browser once the compilation
has completed.