Resource chain demonstration
Enable resource chain handling. Add an extra "dev" profile to disable
the cache (see also gh-3739).

I couldn't manage to demonstrate the difference between this improved
version and the previous one.

NOTE: we should have the dependency management for the whole project
right from the start as it is too easy to forget to add it.