Static resource demonstration
Our problem is that the Javascript code is now located in another module,
and so does the CSS. Spring Boot allows us to keep the same friendly
developer experience.

We need to update our configuration so that Spring Boot looks in those
additional folders: that way, we can edit the source and have immediate
feedback in the application. To make that consistent, we will specify
folders that are relative to the working directory of the application,
that is the `spring-boot-4tw-web` module. We can make sure our local run
configuration uses that by adding `$MODULE_DIR` as working directory in
Intellij IDEA. Build tools are using that directory by default.

Once we've done that, we can update our `dev` profile. Restart can also
be configured to trigger a LiveReload refresh when such file changes.

Now the application process our query in the background we have no visual
indication that it is busy. We can add this based on our setup.