Remote deploy demonstration
We first push our new `boot-versions-provider` to CF and then we can
update the JS live to go to that service instead. Unfortunately, we
"forgot" to add the milestone repo so it only shows the release now.

We start a separate app that watches the file system of this app to send
remote changes to CF. We update `application.properties` to also add the
milestone repo which will redeploy the CF app remotely and we can see now
that the auto-completion works as expected.